Patch Thrift to 0.9.3-p6 to eliminate ssl warnings.
This fixes IMPALA-8333: Remove Impala Shell warnings part 2

Remove misleading deprecation warning from validate method.
This method is used to determine if the peer's SSL certificate should be
validated. The deprecation warning message seems to be the result of a
confusion of the role of the validate method with the use of a
deprecated parameter named validate that is used in the initializer.